# backend/docs/Wulin-OpenApi.yaml
openapi: 3.0.2
info:
  title: 武林书画院项目
  version: SNAPSHOT-1.0.0
  description: 用于设计后台接口以及调试的的文件
  contact:
    name: 陪葬品<FuneralObjects>
    email: funeral_objects@qq.com
    url: https://github.com/jtiiii/wulinacademy
servers:
  - url: http://localhost:9000
    description: 本地开发环境
...
paths:
...
  /folder:
    post:
      summary: 添加文件夹
      tags:
        - folder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/folder-save'
      responses:
        204:
          $ref: '#/components/responses/204'
      security:
        - auth-token: ['stranger']
  /folder/{id}:
    delete:
      summary: 删除文件夹
      tags:
        - folder
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        204:
          $ref: '#/components/responses/204'
      security:
        - auth-token: ['stranger']
  /folder/{id}/son:
    get:
      summary: 查看子文件夹
      tags:
        - folder
      parameters:
        - $ref: '#/components/parameters/id'
      responses:
        200:
          description: 查询成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/folder'
      security:
        - auth-token: ['stranger']
...
components:
  schemas:
    id:
      type: string
      description: 主键Id
    status:
      type: integer
      description: 状态 [VISIBLE(1)/可见，INVISIBLE(0)/ 不可见，DELETED(-1)/ 删除]
      example: 1
...
    createTime:
      type: string
      description: 创建日期
      format: yyyy-MM-dd'T'HH:mm:ss
    updateTime:
      type: string
      description: 更新日期
      format: yyyy-MM-dd'T'HH:mm:ss
...
    folder-save:
      type: object
      properties:
        name:
          type: string
          description: 文件夹名
        parent:
          type: string
          description: 父级id
        user:
          type: string
          description: 所属用户
        status:
          $ref: '#/components/schemas/status'
    folder:
      allOf:
        - $ref: '#/components/schemas/folder-save'
        - type: object
          properties:
            id:
              $ref: '#/components/schemas/id'
            createTime:
              $ref: '#/components/schemas/createTime'
            updateTime:
              $ref: '#/components/schemas/updateTime'
            path:
              type: string
              description: 文件夹路径
...
  parameters:
...
    id:
      name: id
      in: path
      description: 主键ID
      required: true
      schema:
        type: string
...
  securitySchemes:
    auth-token:
      type: apiKey
      description: 安全密钥
      name: auth-token
      in: header
  responses:
    204:
      description: 执行成功，无数据返回
...